+++ CHANGELOG.md
@@ -17,6 +17,8 @@ Changes
 * [UI]: Fixed bug that would send many emails to users when they request a pipeline completion email. (21.05.3)
 * [Developer]: Updated `xlsx` to version 0.17.0
 * [Test]: Updated chromedriver to version 92.0.0.
+* [UI]: Updated Project Single Sample Analyses Outputs pages to use Ant Design.
+* [UI]: Updated Your Single Sample Analysis Outputs page to use Ant Design.
 * [Developer]: Updated redux API within the cart page to use Redux Toolkit.
 
 21.01 to 21.05